# DB-11559 Improve conglomerate cost estimation performance