--- a/db-engine/src/main/java/com/splicemachine/db/iapi/stats/ColumnStatisticsImpl.java
+++ b/db-engine/src/main/java/com/splicemachine/db/iapi/stats/ColumnStatisticsImpl.java
@@ -70,6 +70,8 @@ public class ColumnStatisticsImpl implements ItemStatistics<DataValueDescriptor>
     protected DataValueDescriptor dvd;
     private long rpv=-1; //rows per value excluding skewed values
 
+    private com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] freqSketchNoFpItems;
+
     public ColumnStatisticsImpl() {
 
     }
@@ -151,6 +153,14 @@ public class ColumnStatisticsImpl implements ItemStatistics<DataValueDescriptor>
                 thetaMem.freeMemory();
 
         }
+    }
+
+    private com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] getFreqSketchNoFpItems() {
+        if (freqSketchNoFpItems == null) {
+            freqSketchNoFpItems = frequenciesSketch.getFrequentItems(ErrorType.NO_FALSE_POSITIVES);
+        }
+        assert freqSketchNoFpItems != null : "freqSketchNoFpItems is null";
+        return freqSketchNoFpItems;
     }
 
     /**
@@ -307,7 +317,7 @@ public class ColumnStatisticsImpl implements ItemStatistics<DataValueDescriptor>
             count = frequenciesSketch.getEstimate(lookUpElement);
         } else {
         // Iterated comparing
-            com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] items = frequenciesSketch.getFrequentItems(ErrorType.NO_FALSE_POSITIVES);
+            com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] items = getFreqSketchNoFpItems();
             for (com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor> row: items) {
                 DataValueDescriptor skewedValue = row.getItem();
                 try {
@@ -332,7 +342,7 @@ public class ColumnStatisticsImpl implements ItemStatistics<DataValueDescriptor>
     private long getAvgRowsPerValueExcludingSkews() {
         long skewCount = 0;
         long skewNum = 0;
-        com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] items = frequenciesSketch.getFrequentItems(ErrorType.NO_FALSE_POSITIVES);
+        com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] items = getFreqSketchNoFpItems();
         for (com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor> row: items) {
             skewCount += row.getEstimate();
             skewNum ++;
@@ -348,7 +358,7 @@ public class ColumnStatisticsImpl implements ItemStatistics<DataValueDescriptor>
 
     private long getSkewedRowCountInRange(DataValueDescriptor start, DataValueDescriptor stop, boolean includeStart, boolean includeStop) {
         long skewCount = 0;
-        com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] items = frequenciesSketch.getFrequentItems(ErrorType.NO_FALSE_POSITIVES);
+        com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] items = getFreqSketchNoFpItems();
         for (com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor> row: items) {
             DataValueDescriptor skewedValue = row.getItem();
             try {
@@ -679,7 +689,7 @@ public class ColumnStatisticsImpl implements ItemStatistics<DataValueDescriptor>
     public long selectivityExcludingValueIfSkewed(DataValueDescriptor value) {
         long skewCount = 0;
         long skewNum = 0;
-        com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] items = frequenciesSketch.getFrequentItems(ErrorType.NO_FALSE_POSITIVES);
+        com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor>[] items = getFreqSketchNoFpItems();
         for (com.yahoo.sketches.frequencies.ItemsSketch.Row<DataValueDescriptor> row: items) {
             DataValueDescriptor skewedValue = row.getItem();
             try {