--- a/db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/CastNode.java
+++ b/db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/CastNode.java
@@ -316,17 +316,7 @@ public class CastNode extends ValueNode
             }
         }
 
-        /* We can't chop out cast above an untyped null because
-         * the store can't handle it.
-         */
-        if ((castOperand instanceof ConstantNode) &&
-                !(castOperand instanceof UntypedNullConstantNode))
-        {
-            // Return the new constant if the cast was performed
-            return bindConstantCast();
-        }
-
-        return this;
+        return bindConstantCast();
     }
 
     /**
@@ -337,8 +327,7 @@ public class CastNode extends ValueNode
      *
      * @exception StandardException        Thrown on error
      */
-    public void bindCastNodeOnly()
-            throws StandardException
+    public void bindCastNodeOnly() throws StandardException
     {
 
         /*
@@ -421,6 +410,11 @@ public class CastNode extends ValueNode
                 )
         { setNullability( true ); }
         else { setNullability(castOperand.getTypeServices().isNullable()); }
+    }
+
+    public ValueNode bindImplicitCast() throws StandardException {
+        bindCastNodeOnly();
+        return bindConstantCast();
     }
 
     /**
@@ -1098,6 +1092,13 @@ public class CastNode extends ValueNode
     }
 
     private ValueNode bindConstantCast() throws StandardException {
+        /* We can't chop out cast above an untyped null because
+         * the store can't handle it.
+         */
+        if (!(castOperand instanceof ConstantNode) ||
+                (castOperand instanceof UntypedNullConstantNode)) {
+            return this;
+        }
         /* If the castOperand is a typed constant then we do the cast at
          * bind time and return a constant of the correct type.
          * NOTE: This could return an exception, but we're prepared to